Cloud docs prototyping
This is an experiment to see how we might build two distinct doc sets, one for the current product and one for a managed service. Some pages need to be excluded from the managed service docs, and pages that are included in both sets of docs might need to differ in specific ways.

Summary of this approach:
- Use separate configs for each "product": `_config.yml` for the installed product, and `_config_cloud.yml` for the managed service.
  - In `_config_cloud.yml`:
    - Build the docs to a `/docs/cloud` subdirectory.
    - Use an exclude list to exclude pages. This list is populated during the build (see bullet below on Makefile).
    - Add a `build_for_cloud: true` variable to all pages included in the managed service docs. On pages that need to differ for managed vs. installed, we can use conditional logic to include the right content. Example: v2.1/index.md.
- Customize the Makefile so that running `make cloud` (as opposed to `make serve`) does the following:
  - Executes `scripts/exclude_from_cloud_docs.py`, which scrapes all 2.1 pages and adds any with `cloud: false` in the front-matter to the `exclude` list in `_config_cloud.yml`.
  - Installs dependencies and generates the cloud docs using `_config_cloud.yml`.
- Build the managed service docs with a distinct sidenav. It's just a placeholder right now.
- Make Docs in the topnav a menu.

CI still needs to be adjusted, and no work has been done on the actual docs content. This is PR is intended purely to provoke questions and feedback.